requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.6.0
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.15
django-after-response==0.2.2
django-celery-beat==2.7.0
django-ckeditor==6.7.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-js-asset==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
frozenlist==1.4.1
idna==3.4
kombu==5.4.0
multidict==6.0.5
mysqlclient==2.2.4
pillow==10.4.0
prompt_toolkit==3.0.47
PyJWT==2.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2022.7.1
redis==5.0.8
requests==2.28.2
setuptools==72.1.0
six==1.16.0
sqlparse==0.4.3
tzdata==2024.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
